Move test scripts to scripts folder
