--- test/tag/raw.test.js
@@ -1,9 +1,0 @@
-var should = require('chai').should();
-
-describe('Tag - raw', function(){
-  var raw = require('../../lib/plugins/tag/raw');
-
-  it('content', function(){
-    raw([], '123456789<b>strong</b>987654321').should.eql('123456789<b>strong</b>987654321');
-  });
-});
